requirements.txt
google-adk
google-cloud-aiplatform
python-dotenv
pydantic

